beit2/requirements.txt
torchvision
timm
Pillow
blobfile
mypy
numpy
pytest
requests
einops
tensorboardX
deepspeed
scipy
ftfy
regex
